requirements.txt
flask
python-dotenv
openai==1.23.6
torch==2.0.1
transformers==4.37.2
scikit-learn==1.6.1
pymupdf
pymongo
requests==2.31.0
sentencepiece
